+++ Headers/UXTransitionController.h
@@ -4,10 +4,10 @@
 //     class-dump is Copyright (C) 1997-1998, 2000-2001, 2004-2013 by Steve Nygard.
 //
 
+#import "NSObject.h"
 
-
-#import "UXViewControllerAnimatedTransitioning-Protocol.h"
-#import "UXViewControllerInteractiveTransitioning-Protocol.h"
+#import "UXViewControllerAnimatedTransitioning.h"
+#import "UXViewControllerInteractiveTransitioning.h"
 
 @class NSString;
 
@@ -30,7 +30,7 @@
 // Remaining properties
 @property(readonly, copy) NSString *debugDescription;
 @property(readonly, copy) NSString *description;
-@property(readonly) NSUInteger hash;
+@property(readonly) unsigned long long hash;
 @property(readonly) Class superclass;
 
 @end
