+++ Headers/_UXCollectionView.h
@@ -4,8 +4,9 @@
 //     class-dump is Copyright (C) 1997-1998, 2000-2001, 2004-2013 by Steve Nygard.
 //
 
-#import "UXCollectionView.h"
-#import "_UXCollectionViewOverdraw-Protocol.h"
+#import <UXKit/UXCollectionView.h>
+
+#import "_UXCollectionViewOverdraw.h"
 
 @class NSString;
 
@@ -19,7 +20,7 @@
 // Remaining properties
 @property(readonly, copy) NSString *debugDescription;
 @property(readonly, copy) NSString *description;
-@property(readonly) NSUInteger hash;
+@property(readonly) unsigned long long hash;
 @property(readonly) Class superclass;
 
 @end
